Move IsValidKey to new variableutils package
